+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Whenever `minor` is bumped in the `major.minor.patch` version (e.g. `1.0.2 -> 1.
   - Allow loading an existing campagin with `add-existing`
   - Add a button for downloading campaign sources
   - Move campaign source data storate from `data/tasks/` to `data/campaigns/` with automatic migration
+  - Add global progress tracking to the dashboard
+  - Restyle dashboard
 - v1.1.4
   - Make span toolboxes stick for 0.5 second after selection
 - v1.1.3